refactor(api): move JWT plugin to non-index file
